intelnexo-back/requirements.txt
psycopg2-binary
Flask-SQLAlchemy
blinker==1.6.2
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
PyJWT==2.7.0
PyMySQL==1.0.3
pytest==7.3.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
tomli==2.0.1
Werkzeug==2.3.4
zipp==3.15.0
flask_pydantic